+++ src/features/landing/pricing/hooks/useEducatorCheckoutNavigation.tsx
@@ -5,7 +5,7 @@ import { useRouter } from "next/navigation";
 import { useAuthContext } from "@/app/providers/AuthProvider";
 import { useModal } from "@/app/providers/ModalProvider";
 import { CheckModal } from "@/components/common/modals/CheckModal";
-import { isEducatorRole } from "@/features/landing/lib/educatorRole";
+import { isEducatorRole } from "@/shared/landing/lib/educatorRole";
 
 export function useEducatorCheckoutNavigation() {
   const router = useRouter();
